# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "USC App Degree Projects API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "List Projects",
        "operationId": "list_projects_api_v1_projects_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Status"
            }
          },
          {
            "name": "modality_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Modality Id"
            }
          },
          {
            "name": "academic_period",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Academic Period"
            }
          },
          {
            "name": "academic_program_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Academic Program Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProjectsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "Inscribe una nueva idea de trabajo de grado (solo Estudiante).",
        "operationId": "create_project_api_v1_projects_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/projects/{project_id}/history": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Get Project History",
        "description": "Retorna el historial cronológico (ASC) de todos los eventos del proyecto:\n  1. Cambios de estado (project_status_history)\n  2. Documentos adjuntos (attachments → submissions)\n  3. Calificaciones de jurados (evaluations)\n\nControl de acceso:\n  - Administrador: historial completo de cualquier trabajo\n  - Estudiante: historial de su propio trabajo (sin identidad de jurados)\n  - Docente: historial de trabajos donde tiene función asignada\n\nLos jurados son anónimos para el Estudiante en eventos de calificación.",
        "operationId": "get_project_history_api_v1_projects__project_id__history_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "anyOf": [
                      {
                        "$ref": "#/components/schemas/StatusChangeEvent"
                      },
                      {
                        "$ref": "#/components/schemas/DocumentUploadedEvent"
                      },
                      {
                        "$ref": "#/components/schemas/EvaluationSubmittedEvent"
                      }
                    ]
                  },
                  "title": "Response Get Project History Api V1 Projects  Project Id  History Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DocumentUploadedEvent": {
        "properties": {
          "type": {
            "type": "string",
            "const": "document_uploaded",
            "title": "Type"
          },
          "attachment_type": {
            "type": "string",
            "title": "Attachment Type"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "uploaded_by_name": {
            "type": "string",
            "title": "Uploaded By Name"
          },
          "uploaded_at": {
            "type": "string",
            "format": "date-time",
            "title": "Uploaded At"
          }
        },
        "type": "object",
        "required": [
          "type",
          "attachment_type",
          "file_name",
          "stage",
          "uploaded_by_name",
          "uploaded_at"
        ],
        "title": "DocumentUploadedEvent",
        "example": {
          "attachment_type": "plantilla",
          "file_name": "anteproyecto_v1.pdf",
          "stage": "anteproyecto",
          "type": "document_uploaded",
          "uploaded_at": "2026-02-20T16:05:00Z",
          "uploaded_by_name": "Ana María Torres"
        }
      },
...
      "EvaluationSubmittedEvent": {
        "properties": {
          "type": {
            "type": "string",
            "const": "evaluation_submitted",
            "title": "Type"
          },
          "juror_number": {
            "type": "integer",
            "title": "Juror Number"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "is_extemporaneous": {
            "type": "boolean",
            "title": "Is Extemporaneous"
          },
          "juror_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Juror Name"
          }
        },
        "type": "object",
        "required": [
          "type",
          "juror_number",
          "score",
          "stage",
          "submitted_at",
          "is_extemporaneous"
        ],
        "title": "EvaluationSubmittedEvent",
        "example": {
          "is_extemporaneous": false,
          "juror_name": "Dra. Laura Martínez",
          "juror_number": 1,
          "score": 3.8,
          "stage": "anteproyecto",
          "submitted_at": "2026-03-01T14:00:00Z",
          "type": "evaluation_submitted"
        }
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginatedProjectsResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ProjectResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "PaginatedProjectsResponse",
        "example": {
          "items": [],
          "page": 1,
          "size": 20,
          "total": 0
        }
      },
...
      "ProjectCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 100,
            "title": "Title"
          },
          "modality_id": {
            "type": "string",
            "format": "uuid",
            "title": "Modality Id"
          },
          "academic_program_id": {
            "type": "string",
            "format": "uuid",
            "title": "Academic Program Id"
          },
          "research_group": {
            "type": "string",
            "title": "Research Group"
          },
          "research_line": {
            "type": "string",
            "maxLength": 200,
            "title": "Research Line"
          },
          "suggested_director": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 150
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Director"
          },
          "member_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "minItems": 1,
            "title": "Member Ids"
          },
          "prerequisite_declaration": {
            "type": "boolean",
            "title": "Prerequisite Declaration"
          }
        },
        "type": "object",
        "required": [
          "title",
          "modality_id",
          "academic_program_id",
          "research_group",
          "research_line",
          "member_ids",
          "prerequisite_declaration"
        ],
        "title": "ProjectCreate",
        "example": {
          "academic_program_id": "a1b2c3d4-1234-5678-abcd-ef0123456789",
          "member_ids": [
            "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          ],
          "modality_id": "b2c3d4e5-2345-6789-bcde-f01234567890",
          "prerequisite_declaration": true,
          "research_group": "GIEIAM",
          "research_line": "Ingeniería de software y sistemas de información",
          "suggested_director": "Dr. Carlos Rodríguez",
          "title": "Sistema de gestión de trabajos de grado USC"
        }
      },
...
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "modality_id": {
            "type": "string",
            "format": "uuid",
            "title": "Modality Id"
          },
          "academic_program_id": {
            "type": "string",
            "format": "uuid",
            "title": "Academic Program Id"
          },
          "research_group": {
            "type": "string",
            "title": "Research Group"
          },
          "research_line": {
            "type": "string",
            "title": "Research Line"
          },
          "suggested_director": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Suggested Director"
          },
          "period": {
            "type": "string",
            "title": "Period"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "has_company_link": {
            "type": "boolean",
            "title": "Has Company Link"
          },
          "plagiarism_suspended": {
            "type": "boolean",
            "title": "Plagiarism Suspended"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "modality_id",
          "academic_program_id",
          "research_group",
          "research_line",
          "suggested_director",
          "period",
          "status",
          "has_company_link",
          "plagiarism_suspended",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectResponse",
        "example": {
          "academic_program_id": "a1b2c3d4-1234-5678-abcd-ef0123456789",
          "created_at": "2026-02-05T09:00:00Z",
          "has_company_link": false,
          "id": "f6a7b8c9-6789-abcd-f012-345678901234",
          "modality_id": "b2c3d4e5-2345-6789-bcde-f01234567890",
          "period": "2026-1",
          "plagiarism_suspended": false,
          "research_group": "GIEIAM",
          "research_line": "Ingeniería de software y sistemas de información",
          "status": "idea_inscrita",
          "suggested_director": "Dr. Carlos Rodríguez",
          "title": "Sistema de gestión de trabajos de grado USC",
          "updated_at": "2026-02-05T09:00:00Z"
        }
      },
...
      "StatusChangeEvent": {
        "properties": {
          "type": {
            "type": "string",
            "const": "status_change",
            "title": "Type"
          },
          "previous_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Previous Status"
          },
          "new_status": {
            "type": "string",
            "title": "New Status"
          },
          "changed_by_name": {
            "type": "string",
            "title": "Changed By Name"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "changed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Changed At"
          }
        },
        "type": "object",
        "required": [
          "type",
          "previous_status",
          "new_status",
          "changed_by_name",
          "reason",
          "changed_at"
        ],
        "title": "StatusChangeEvent",
        "example": {
          "changed_at": "2026-02-08T11:00:00Z",
          "changed_by_name": "Admin USC",
          "new_status": "idea_aprobada",
          "previous_status": "idea_inscrita",
          "reason": "Cumple con todos los requisitos.",
          "type": "status_change"
        }
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}